+++ xash3d_engine/engine/client/s_vox.c
@@ -515,7 +515,8 @@ static void Test_VOX_GetDirectory(void)
 		{"", "", "vox/", "bark bark", "bark bark", "vox/", "barney/meow", "meow", "barney/"
 
 		};
-	int i;
+
+	size_t i;
 
 	for ( i = 0; i < sizeof(data) / sizeof(data[0]); i += 3 )
 	{
@@ -529,7 +530,7 @@ static void Test_VOX_GetDirectory(void)
 
 static void Test_VOX_LookupString(void)
 {
-	int i;
+	size_t i;
 	const char *p,
 		*data[] = {
 			"0",
@@ -588,7 +589,8 @@ static void Test_VOX_ParseString(void)
 		"_period",
 		NULL,
 	};
-	int i = 0;
+
+	size_t i = 0;
 
 	while ( i < sizeof(data) / sizeof(data[0]) )
 	{
